refactor(rename): useTabs 변수 useMemo 수정 및 폴더 경로 수정
