+++ src/components/ui/tabs/TabsPill.tsx
@@ -2,7 +2,7 @@
 
 import { useMemo } from 'react';
 import { cn } from '@/lib/cn';
-import { useTabs } from './useTabs';
+import { useTabs } from './hooks/useTabs';
 import type { TabItem } from './utils/tabIds';
 
 interface TabsPillProps<T extends string> {
