+++ src/components/ui/tabs/index.ts
@@ -1,4 +1,4 @@
-export { useTabs } from './useTabs';
+export { useTabs } from './hooks/useTabs';
 export { buildTabId, buildPanelId, type TabItem } from './utils/tabIds';
 export { TabsUnderline } from './TabsUnderline';
 export { TabsPill } from './TabsPill';
